requirements.txt
Django
django-debug-toolbar
python-decouple
pytz
sqlparse
django-rest-knox
djangorestframework
channels
channels_redis
django-rest-framework-social-oauth2